requirements.txt
numpy<2.0.0,>=1.23.5
pillow==10.4.0
streamlit==1.37.0
tensorflow==2.17.0

